Potential fix for dropped pulse on peerless start
This is a potential bugfix for the missing initial clock pulse when
pushing play without any Link peers. The ideal behavior is for the clock
to start immediately rather than waiting for the next shared timeline
downbeat.

Previously, the outpuloop was resetting the timeline if Link
was cued, immediately before switching to Playing mode. However, all of
the timeline beat/phase/time state capture takes place earlier in the
loop than the playback state logic, so at that point the loop is working
with old data.

In addition to refactoring so the timeline reset takes place in the
button press handler, I also changed it to request beat zero 1ms in the
future in hopes of giving the output loop enough "head start" to pick up
the beat 0 downbeat.