Point bin/dockcross script at our own dockerhub fork
Point bin/dockcross at circuithappy/cross-linux-armv7 which is a fork
fork of the original whiskeybravo/dockcross-linux-armv7 image. They are
otherwise identical. The image is an updated version of the dockcross
armv7 image based on stretch instead of jessie, to ensure we have the
necessary minimum gcc version to build the ableton link library.